backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0.36
pydantic
pydantic-settings
python-multipart
cors
passlib
bcrypt
python-jose
google-generativeai
psycopg2-binary
